// server/src/user/enums/user-error.enum.ts
export enum UserError {
    USER_NOT_FOUND = 'USER_NOT_FOUND',

    USER_ID_EMPTY = 'USER_ID_EMPTY',

    USER_NAME_EMPTY = 'USER_NAME_EMPTY',
    USER_NAME_EXISTED = 'USER_NAME_EXISTED',
    USER_NAME_NOT_EXISTS = 'USER_NAME_NOT_EXISTS',

    FULL_NAME_EMPTY = 'FULL_NAME_EMPTY',

    PHONE_NUMBER_EMPTY = 'PHONE_NUMBER_EMPTY',
    PHONE_NUMBER_INVALID = 'PHONE_NUMBER_INVALID',
    PHONE_NUMBER_EXISTED = 'PHONE_NUMBER_EXISTED',
    PHONE_NUMBER_NOT_EXISTS = 'PHONE_NUMBER_NOT_EXISTS',

    NEW_PHONE_NUMBER_EMPTY = 'NEW_PHONE_NUMBER_EMPTY',
    NEW_PHONE_NUMBER_INVALID = 'NEW_PHONE_NUMBER_INVALID',

    EMAIL_EMPTY = 'EMAIL_EMPTY',
    EMAIL_INVALID = 'EMAIL_INVALID',
    EMAIL_EXISTED = 'EMAIL_EXISTED',
    EMAIL_NOT_EXISTS = 'EMAIL_NOT_EXISTS',

    NEW_EMAIL_EMPTY = 'NEW_EMAIL_EMPTY',
    NEW_EMAIL_INVALID = 'NEW_EMAIL_INVALID',

    PASSWORD_EMPTY = 'PASSWORD_EMPTY',
    PASSWORD_NOT_CORRECT = 'PASSWORD_NOT_CORRECT',
    PASSWORD_LENGTH = 'PASSWORD_LENGTH',
    PASSWORD_INVALID = 'PASSWORD_INVALID',

    NEW_PASSWORD_EMPTY = 'NEW_PASSWORD_EMPTY',
    NEW_PASSWORD_LENGTH = 'NEW_PASSWORD_LENGTH',
    NEW_PASSWORD_INVALID = 'NEW_PASSWORD_INVALID',

    CONFIRM_PASSWORD_EMPTY = 'CONFIRM_PASSWORD_EMPTY',
    CONFIRM_PASSWORD_NOT_MATCH = 'CONFIRM_PASSWORD_NOT_MATCH',

    GENDER_EMPTY = 'GENDER_EMPTY',
    GENDER_INVALID = 'GENDER_INVALID',

    DOB_EMPTY = 'DOB_EMPTY',
    DOB_INVALID = 'DOB_INVALID',
}